feat: added image and gif support
